[PDS-152] refactor: car와 resource의 상위 product 패키지 생성
