+++ src/main/java/com/example/pladialmserver/booking/repository/resourceBooking/ResourceBookingRepositoryImpl.java
@@ -4,8 +4,8 @@ import com.example.pladialmserver.booking.dto.response.BookingRes;
 import com.example.pladialmserver.booking.entity.ResourceBooking;
 import com.example.pladialmserver.global.entity.BookingStatus;
 import com.example.pladialmserver.global.utils.DateTimeUtil;
-import com.example.pladialmserver.resource.dto.response.ResourceBookingRes;
-import com.example.pladialmserver.resource.entity.Resource;
+import com.example.pladialmserver.product.resource.dto.response.ResourceBookingRes;
+import com.example.pladialmserver.product.resource.entity.Resource;
 import com.example.pladialmserver.user.entity.User;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
