--- src/main/java/com/example/pladialmserver/car/repository/CarCustom.java
@@ -1,15 +1,0 @@
-package com.example.pladialmserver.car.repository;
-
-import com.example.pladialmserver.car.dto.CarRes;
-import com.example.pladialmserver.car.entity.Car;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-import java.time.LocalDate;
-
-public interface CarCustom {
-//    Page<CarRes> findAvailableCars(String carName, LocalDate startDate, LocalDate endDate, Pageable pageable);
-
-//    Page<CarRes> findAllCars(Pageable pageable);
-
-}
