+++ src/main/java/com/example/pladialmserver/office/controller/OfficeAdminController.java
@@ -4,7 +4,7 @@ import com.example.pladialmserver.global.resolver.Account;
 import com.example.pladialmserver.global.response.ResponseCustom;
 import com.example.pladialmserver.office.dto.response.AdminOfficesDetailsRes;
 import com.example.pladialmserver.office.service.OfficeService;
-import com.example.pladialmserver.resource.dto.request.CreateOfficeReq;
+import com.example.pladialmserver.product.resource.dto.request.CreateOfficeReq;
 import com.example.pladialmserver.user.entity.User;
 import io.swagger.annotations.Api;
 import io.swagger.v3.oas.annotations.Operation;
