+++ src/main/java/com/example/pladialmserver/product/car/repository/CarCustom.java
@@ -1,0 +1,8 @@
+package com.example.pladialmserver.product.car.repository;
+
+public interface CarCustom {
+//    Page<CarRes> findAvailableCars(String carName, LocalDate startDate, LocalDate endDate, Pageable pageable);
+
+//    Page<CarRes> findAllCars(Pageable pageable);
+
+}
