+++ src/main/java/com/example/pladialmserver/product/resource/dto/response/ResourceBookingRes.java
@@ -1,0 +1,32 @@
+package com.example.pladialmserver.product.resource.dto.response;
+
+import com.example.pladialmserver.booking.entity.ResourceBooking;
+import com.example.pladialmserver.global.utils.DateTimeUtil;
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ResourceBookingRes {
+    @Schema(type = "String", description = "예약 시작 시간", example = "2023-10-23 13:00")
+    private String startDateTime;
+    @Schema(type = "String", description = "예약 종료 시간", example = "2023-10-24 15:00")
+    private String endDateTime;
+    @Schema(type = "String", description = "예약자명", example = "소징")
+    private String reservatorName;
+    @Schema(type = "String", description = "예약자 부서", example = "미디어")
+    private String reservatorDepartment;
+    @Schema(type = "String", description = "예약자 연락처", example = "010-1212-1111")
+    private String reservatorPhone;
+
+    public static ResourceBookingRes toDto(ResourceBooking booking) {
+        return ResourceBookingRes.builder()
+                .startDateTime(DateTimeUtil.dateTimeToString(booking.getStartDate()))
+                .endDateTime(DateTimeUtil.dateTimeToString(booking.getEndDate()))
+                .reservatorName(booking.getUser().getName())
+                .reservatorDepartment(booking.getUser().getDepartment().getName())
+                .reservatorPhone(booking.getUser().getPhone())
+                .build();
+    }
+}
