+++ src/main/java/com/example/pladialmserver/product/resource/dto/response/ResourceRes.java
@@ -1,0 +1,34 @@
+package com.example.pladialmserver.product.resource.dto.response;
+
+import com.example.pladialmserver.global.utils.AwsS3ImageUrlUtil;
+import com.example.pladialmserver.product.resource.entity.Resource;
+import lombok.Builder;
+import lombok.Data;
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Data
+@Builder
+public class ResourceRes {
+    @Schema(type = "Long", description = "장비 Id", example = "1")
+    private Long resourceId;
+    @Schema(type = "String", description = "이미지 url", example = "'img.key'")
+    private String imgUrl;
+    @Schema(type = "String", description = "장비 이름", example = "'벤츠'")
+    private String name;
+    @Schema(type = "String", description = "보관장소", example = "'4층 창고'")
+    private String location;
+    @Schema(type = "String", description = "설명", example = "'승학이 보물'")
+    private String description;
+
+    // TODO 기획 변경으로 인한 수정
+    public static ResourceRes toDto(Resource resource){
+        return ResourceRes.builder()
+                .resourceId(resource.getResourceId())
+                .imgUrl(AwsS3ImageUrlUtil.toUrl(resource.getImgKey()))
+                .name(resource.getName())
+                .location(resource.getLocation())
+                .description(resource.getDescription())
+                .build();
+    }
+
+}
