+++ src/main/java/com/example/pladialmserver/product/resource/dto/response/ResourcesList.java
@@ -1,6 +1,6 @@
-package com.example.pladialmserver.office.dto.response;
+package com.example.pladialmserver.product.resource.dto.response;
 
-import com.example.pladialmserver.booking.entity.OfficeBooking;
+import com.example.pladialmserver.booking.entity.ResourceBooking;
 import com.example.pladialmserver.global.utils.DateTimeUtil;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
@@ -8,8 +8,8 @@ import lombok.Data;
 
 @Data
 @Builder
-public class OfficesList {
-    @Schema(type="String", description="예약자 이름", example ="이승학")
+public class ResourcesList {
+    @Schema(type = "String", description = "예약자 이름", example = "이승학")
     private String reservatorName;
     @Schema(type = "String", description = "예약자 전화번호", example = "010-0000-0000")
     private String reservatorPhone;
@@ -23,14 +23,14 @@ public class OfficesList {
     private String bookingStatus;
 
     // TODO 기획 변경으로 인한 수정
-    public static OfficesList toDto(OfficeBooking officeBooking){
-        return OfficesList.builder()
-                .reservatorName(officeBooking.getUser().getName())
-                .reservatorPhone(officeBooking.getUser().getPhone())
-                .startDateTime(DateTimeUtil.dateAndTimeToString(officeBooking.getDate(),officeBooking.getStartTime()))
-                .endDateTime(DateTimeUtil.dateAndTimeToString(officeBooking.getDate(),officeBooking.getEndTime()))
-                .goal(officeBooking.getMemo())
-                .bookingStatus(officeBooking.getStatus().getValue())
+    public static ResourcesList toDto(ResourceBooking resourceBooking){
+        return ResourcesList.builder()
+                .reservatorName(resourceBooking.getUser().getName())
+                .reservatorPhone(resourceBooking.getUser().getPhone())
+                .startDateTime(DateTimeUtil.dateTimeToString(resourceBooking.getStartDate()))
+                .endDateTime(DateTimeUtil.dateTimeToString(resourceBooking.getEndDate()))
+                .goal(resourceBooking.getMemo())
+                .bookingStatus(resourceBooking.getStatus().getValue())
                 .build();
     }
 }
