--- src/main/java/com/example/pladialmserver/resource/repository/ResourceCustom.java
@@ -1,9 +1,0 @@
-package com.example.pladialmserver.resource.repository;
-
-import com.example.pladialmserver.resource.dto.response.AdminResourcesRes;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface ResourceCustom {
-    Page<AdminResourcesRes> search(String keyword, Pageable pageable);
-}
